Move different users into seperate files
